+++ integreat_cms/api/v3/locations.py
@@ -90,6 +90,7 @@ def transform_poi_translation(poi_translation):
         "opening_hours": opening_hours,
         "location": transform_poi(poi),
         "hash": None,
+        "stops": poi.stops or None,
     }