+++ CHANGELOG.md
@@ -1,6 +1,8 @@
 UNRELEASED
 ----------
 
+* [ [#959](https://github.com/digitalfabrik/integreat-cms/issues/959) ] Add stops in the vicinities to POI
+
 
 2023.1.2
 --------